simOCR/requirements.txt
Flask==1.0.2
opencv_python==4.1.0.25
Werkzeug==0.14.1
matplotlib==3.0.3
pytesseract==0.2.7
numpy==1.16.2
kraken==2.0.5
imutils==0.5.2
Pillow==6.1.0
